+++ pywal/backends/wal.py
@@ -9,7 +9,6 @@ import subprocess
 import sys
 
 from .. import colors
-from .. import util
 
 
 def imagemagick(color_count, img, magick_command):
@@ -26,12 +25,12 @@ def imagemagick(color_count, img, magick_command):
 
     try:
         output = subprocess.check_output(
-            [*magick_command, img, *flags], stderr=subprocess.STDOUT
+            [*magick_command.split(), img, *flags], stderr=subprocess.STDOUT
         ).splitlines()
     except subprocess.CalledProcessError as Err:
         logging.error("Imagemagick error: %s", Err)
         logging.error(
-            "IM 7 disables stdout by default, check the wiki for the fix."
+            "IM 7 disables stdout by default, check the manual or wiki to fix."
         )
         sys.exit(1)
     return output
@@ -39,53 +38,87 @@ def imagemagick(color_count, img, magick_command):
 
 def has_im():
     """Check to see if the user has im installed."""
+    magick_commands = []
+
     if shutil.which("magick"):
-        return ["magick", "convert"]
+        magick_commands.extend(["magick", "magick convert"])
 
     if shutil.which("convert"):
-        return ["convert"]
+        magick_commands.append("convert")
+
+    if len(magick_commands) > 0:
+        return magick_commands
 
     logging.error("Imagemagick wasn't found on your system.")
     logging.error("Try another backend. (wal --backend)")
     sys.exit(1)
 
 
-def try_gen_in_range(img, magick_command):
-    for i in range(0, 20, 1):
-        raw_colors = imagemagick(16 + i, img, magick_command)
+def gen_colors_with_command(
+        img, magick_command, beginning_color_count=16, iteration_count=20
+        ):
+    """Iteratively attempt to generate a 16-color palette
+    using a specific Imagemagick command."""
+    hex_pattern = re.compile(r"#[A-F0-9]{6}", re.IGNORECASE)
 
-        if len(raw_colors) > 16:
+    max_color_count = beginning_color_count + iteration_count - 1
+    for color_count in range(
+            beginning_color_count, beginning_color_count + iteration_count
+            ):
+        raw_output = imagemagick(color_count, img, magick_command)
+        hex_colors = [
+            hex_pattern.search(str(col)).group()
+            for col in raw_output
+            if hex_pattern.search(str(col))
+        ]
+
+        if len(hex_colors) >= 16:
             break
 
-        if i == 19:
-            logging.error("Imagemagick couldn't generate a suitable palette.")
-            sys.exit(1)
-
+        if color_count < max_color_count:
+            logging.warning(
+                    "Imagemagick couldn't generate a "
+                    f"palette with {magick_command}."
+                    )
+            logging.warning(
+                    f"Trying a larger palette size {color_count}."
+                    )
         else:
-            logging.warning("Imagemagick couldn't generate a palette.")
-            logging.warning("Trying a larger palette size %s", 16 + i)
-    return raw_colors
+            logging.error(
+                    "Imagemagick couldn't generate a suitable palette "
+                    f"with {magick_command}."
+                    )
+            logging.warning(
+                    "Will try to do palette concatenation, "
+                    "good results not guaranteed!"
+                    )
+            while len(hex_colors) < 16:
+                hex_colors.extend(hex_colors)
+    return hex_colors
 
 
 def gen_colors(img):
-    """Format the output from imagemagick into a list
-    of hex colors."""
-    magick_command = has_im()
+    """Try each Imagemagick command until a color palette
+    is successfully generated."""
+    magick_commands = has_im()
 
-    raw_colors = try_gen_in_range(img, magick_command)
+    for magick_command in magick_commands:
+        logging.debug(f"Trying {magick_command}...")
 
-    try:
-        out = [re.search("#.{6}", str(col)).group(0) for col in raw_colors[1:]]
-    except AttributeError:
-        if magick_command == ["magick", "convert"]:
-            logging.warning("magick convert failed, using only magick")
-            magick_command = ["magick"]
-            raw_colors = try_gen_in_range(img, magick_command)
-            out = [
-                re.search("#.{6}", str(col)).group(0) for col in raw_colors[1:]
-            ]
+        hex_colors = gen_colors_with_command(img, magick_command)
 
-    return out
+        if not hex_colors:
+            logging.warning(
+                    f"Failed to generate colors with {magick_command}."
+                    )
+            continue
+
+        return hex_colors
+
+    raise RuntimeError(
+            "Failed to generate color palette from "
+            f"{img} with these commands: {magick_commands}"
+            )
 
 
 def adjust(cols, light, **kwargs):