+++ pywal/settings.py
@@ -13,7 +13,7 @@ import os
 import platform
 
 
-__version__ = "3.8.6"
+__version__ = "3.8.14"
 __cache_version__ = "2.0.0"